Fixed bug 10199: Most transcoding doesn't work with non-ascii-characters in filename Ensure that the building of the transcoding command is done in 'use bytes' mode so that native byte-strings are used throughout and one does not get accidental promotion to a decoded UFT-8 string (see http://perldoc.perl.org/perlunicode.html).
This would only happen if one of the inputs had the Perl utf8 flag set and, for some strange reason,
the $quality (from the lameQuality preference) variable does have this flag (even though the value is "9").
But there could be other cases with custom convert files or unusual system configurations where
some of the paths might include non-ASCII characters. 

Bug: 10199
git-svn-id: file:///home/awy/mirror/slim/7.6/trunk/server@31416 62299810-d8cb-41fd-93b7-d32162e5a4a4