Fixed bug 16830: CLI returns double-encoded UTF-8 for metadata when local not using UTF-8 encoding Fixes for the original problem: Fix encoding of simple values in Slim::Control::Request::renderAsArray(). All input to renderAsArray() should already be in perl internal character-string format, so just use Slim::Utils::Unicode::utf8encode() instead of Slim::Utils::Unicode::from_to().
Fixes to ensure that the CLI charset param (continues to ?) work with the changes above.
Slim::Control::Stdio::array_to_string() needs to allow output character-set encodings other than just UTF-8.
The current locale has nothing to do with encoding CLI results.

Fixes to related issues discovered as part of the investigation:
Slim::Utils::Unicode::encodingFromString() will return 'utf8' immediately if the flag is already set on the supplied string, unless this test is defeated by a parameter. This will avoid the problem of double-encoding when the encoding is already known.
utf8::decode() is not idempotent so protect against this in a few places.


Bug: 16830
git-svn-id: file:///home/awy/mirror/slim/7.6/trunk/server@31847 62299810-d8cb-41fd-93b7-d32162e5a4a4