BMF caching would store cleaned up filenames instead of original file names. Thus reading from the cache might lead to invalid results, as eg. links (*.lnk) would be represented without their file extension.
Don't cache the modified item list, but the original as read from the disk. Use Tie::Cache::LRU::Expires to cache max. 5 levels of folders for a short period of time (15s).

I verified this change using shortcuts to folders on the same disk, a removable disk and an UNC path, using web UI, Boom and Touch.

Bug: 17177
git-svn-id: file:///home/awy/mirror/slim/7.6/branches/onebrowser/server@32499 62299810-d8cb-41fd-93b7-d32162e5a4a4