Unnecessary items in context menu for current-playlist tracks
Take 2: use place-holder items in feed result that are not included in actual menus because subsequent XMLBrowser navigation invocations may not include the same context parameters as the original menu call.

‪1.‬ No "Remove from playlist" if only item in playlist
‪2.‬ No "Play Next" if only item in playlist or current (playing) item or next item
‪3.‬ No "Play" if current item and playing.

When playing a radio station this removes 3 unnecessary items from the top of the CM. It also removes some unnecessary items in other circumstances. 

git-svn-id: file:///home/awy/mirror/slim/7.7/trunk/server@33858 62299810-d8cb-41fd-93b7-d32162e5a4a4