Resolve a couple of problems player-status subscription reliability.
Need to extract clientId from /slim/unsubscribe Comet requests. This was one of the causes of SP clients not always reconnecting "completely". If the first request in a packet was an unsubscribe then the whole packey would be dropped.

Do not remove the repeating subscription from a (player)status query when the client has gone as otherwise a Comet subscription can be dropped.

Do not give up processing requests from a Comet packet upon any error, other than conditions that explicitly do this.

Stop processing requests from a Comet packet after getting a subscription request that results in an errorNeedClient condition.

Remove duplicate checking for valid clientId.


git-svn-id: file:///home/awy/mirror/slim/7.8/trunk/server@33928 62299810-d8cb-41fd-93b7-d32162e5a4a4