Enable non-support of ip3k players (first pass).
This change uses the constant main::IP3K to support removal of
support for ip3k-ased players. This primarily means the Buttons
and IR handling for those players and their graphic displays.

The change is incomplete on at least two grounds:
1. The CLI "button" command is used by SqueezePlay UIs for a number
of functions includng presetPlay, repeatToggle. This functionality
relies on both the IR and Buttons subsystems. It needs to be rewritten.
2. Lots of places make use of Slim::Buttons and not all of these have
been protected (especially in Plugins).

Change-Id: Ifaa99ad4fd4a9b2406f9ed80a634878f9f19a491