Map Content-type for AAC and PCM when downloading.
When using the /music/id/download mechanism to stream a file,
as is used in UEML, some formats need to be mapped from their
file format to their stream format in order to get the correct
Content-Type header. Without this, the server may reject the
stream, when it receives the HTTP headers, as the content-type
is not supported by the player.

Change-Id: Ieb81c6ba0c844fd6b89b885c19a1e49f882a5e7c