Remove non-standard handling of request callback for playlistXitem.
There was code, which appears to have been a legacy from before
the notification mechanism was introduced (although, it seems
unlikely that this is a full explanation), which handled calling
a request's callback function (if one is set) directly within the
code dealing with playlistXitemCommand() instead of leaving it
be be handled by requestDone. A search has revealed no (remaining?)
places where playlistXitemCommand() would be invoked with a callback
function set, let alone where the ordering changes effected by the
non-standrad handling would have any effect.

Change-Id: I4f28a85c5d3110a584b8035b79392f7e62b6ea16