Pass any error string from SN for delegated playlist to caller.
In order that any error that may arise in passing a playlist-
modification request to SN from UEML may be returned to the
invoking UI for display to the user,
Slim::Player::Playlist::addTracks() is extended to take a $callback
parameter. This is invoked upon completion of the function,
possibly asynchrously. The only parameter to the callback is the
error-message string, if there was one.

In order to ensure that the error above is propagated back to the
original request, the various mutual request calls between
playlistcontrolCommand(), playlistXtracksCommand(), playlistXalbumCommand()
and playlistXitemCommand() need to use the callback mechanism
of sub-requests.

Change-Id: I0abf9e89e34878c7fefe3665354b0139fe56d89f