Refactor button-remapping code.
Put button remapping in playerStatus.buttons instead of
playerStatus.remoteMeta.buttons. Only put it there, and not also in the
individual song entries in the playlist loop.

Use 'like' and 'dislike' for the relevant button entries created by
some protocol handlers, and disable the 'repeat' and 'shuffle' buttons
as appropriate.

Use a button remapping of type 'like' for the repeat button and one of
'dislike' for the shuffle button in the Web-UI.

TTP: TSK-2352, Story-225
Change-Id: If89e10a43d0261ac5e47715cf31e074453001004