Do not persist StreamingController TRACKWAIT state
Presisting StreamingController state is only done so that it can be
restored in SlimService in another instance. The TRACKWAIT state is used
when asynchronous callback of a closure is expected and such state cannot
be meaningfully restored.

Rely on subsequent event (possibly resent by player after reconnect)
to trigger next action

Change-Id: I2cd85e18b34302589e39c4366e1be66a8cd9163b