Make Play the default action when browsing favourites.
The Favorties plugin cliBrowse() method passing the special '__goIsPlay'
parameter to XmlBrowser. This causes XmlBrowser to set the base action
(in menu mode) for 'go' to be the same as for 'play'.

The result is for Slimbrowse-based UIs, the default action for
favourite items of type=playlist is to play them rather than to
browse into them.

Non-playable favourites items (if it is possible to create them) will still be
browsed as the default action.

Change-Id: I42f35795acf623f25e5d85e20c55752a6ff3694b