Make a variety of RandomPlay work with UEML
Make the default number of tracks to be added 100 (instead of 10)
for disconnected players.

For disconnected players, always try to add as many tracks as specified
by the addTracks preference (100, above) even when in album or artist mode.
This could result in adding a partial album (or only some of an artist's songs).

For disconnected players, the playlist will not be refilled as it starts
to empty, but with 100 tracks added at a time, that is probably not
a major issue.

Provide a single menu item for random play from the plugin to Slimbrowse
clients. The old code used to provide a node and each of the items below
that, which is unnecessarily complex and did not work with Belsonic.

Extend the idea that a genre CM may provide random play
of tracks or albums from that genre, to artist and year CMs too.

Change-Id: I5d5cbe7a6ceca7f16126fca1a376e6a36ceed1bb