Don't worry about the file extension of media file softlink targets
SlimServer allows softlinks within the media library. In the case of
files that are softlinks it currently ensures that both the softlink
and the softlink target have file extensions that suggest that the
file is a valid media file.

This is generally OK, but I'm hoping to keep my music in 'git
annex' (http://git-annex.branchable.com) and it replaces media files
with a shortcut of the same name to a file held deep in the .git
folder structure. The link target, however, loses its file extension.
In turn this means SlimServer rejects these media files because it
insists that the file extension of both the shortcut and the shortcut
target is one of the enabled file extensions.

Whist I can see the value in checking the link target in the case
where the link points to something totally bogus, it seems unlikely
that a shortcut with the correct file extension would be created
within the music library hierarchy and that target file wasn't a music
file that you intended SlimServer to see.

This commit removes the check of the extension of the shortcut target,
but keeps the check for normal media files and the shortcut itself.