Fix date/time display preference on ip3k players
We have some mix of all lowercase timeformat and mixed case timeFormat
preferences on mysb.com. As MySQL treats them the same, but the caching
level does not, some changes never make it to the database, thus are
reset at every player reconnection, or, worse, don't apply at all (when
doen through the web UI).

Add a clientMigrate handler to _remove_ the old preference before
adding it back with the correct name.