Prevent parallel lookups of the same data.
Set the fetchingMeta flag asap to prevent multiple parallel lookups of
the same data. If that flag is not set, calls for all the tracks in a
playlist could trigger the full request for each of them, as the flag is
only set once the first request has handled all tracks in the playlist.