Bug 17571 (and many more dupes) LMS would fail any DNS lookup if network interface got re-configured during run time, or if it wasn't ready at startup.
The resolver only read the network configuration when the very first DNS
lookup was due. If the configuration changed (because the network only
got setup after LMS started, or because the connection changed), that
invalid configuration would be re-used, leading to DNS failures. 

This change will reset the resolver configuration if a well known domain
lookup such as mysqueezebox.com or radiotime.com occurs. This should
force a re-initialization of the resolver.

Please note that this can still take a few minutes to recover. But it
should no longer require a server restart.