More BMF tweaking
When we first read a folder's content, we only want to know whether an
item should be shown or not. We don't need to create track objects at
this stage yet, as we're going to page through the items later. In the
case of a folder with a lot of items in it we would have created a track
object for every single item, though we probably only need the first
fifty (web UI) or 200 (player UI) items. 

Speed up initial menu creation by only checking for the content type,
without creating the track objects in the database. Only create them as
we're paging through the full result list. This should return control to
the player much quicker, with less risk of interrupting playback, timing
out etc.