Add new Extended Browse Modes plugin to support browsing for Album Artist, Composer etc.
This is a non-working initial checkin. It includes some small changes to
the core code to support filtering by role_id, as well as a simple POC
plugin to add some more browse modes. This plugin can be used as is as
well as extended by 3rd party devs to add even more or more complex
modes.