Add initial support for multiple libraris
- new table library_track stores relation of tracks and libraries
- albums/titles/genres etc. queries can filter by library

A plugin "LibraryDemo" would create a virtual library for tracks longer
than 10 minutes. It's rather simplistic, but shows two ways to get the
library values in to the db (one for the standalone scanner, one for the
LMS internal scanner). Two menu items for albums and artists would be
added to My Music.