Add support for global virtual library handling. BrowseModes would read a pref library_id whenever called to limit results to subsets of the complete library.
The library_id can be a per-player pref or globally set. There's no code
yet to set any of this values.

Please note that you'll need to DELETE the library.db file and run a
full rescan for this change to work. There has been another schema
change I don't want to create a new schema definition file for. You're
living on the edge ;-).