Play local files without storing them in the database
There are various use cases where we would like to play local files
without them being stored in the database:
- BMF while a scan is running
- playing files from removable media (eg. USB sticks)

In order to allow this, we'll have to make sure LMS doesn't treat such
files the same way as regular files. Create ProtocolHandler class based
on File which deals with tmp:// file type. Modify the musicfolder query
to use this URL scheme instead of file:// when needed (or on request).