Bug 9687 - Add option to disable automatically triggered library scans
Add new preference dontTriggerScanOnPrefChange. Pref change handlers
triggering a wipecache would only add their request to the queue, when
dontTriggerScanOnPrefChange is set. The user is prompted to rescan his
library as long as this scan has not been run.

dontTriggerScanOnPrefChange is enabled by default, except on SqueezeOS,
where we don't have a web UI to ask the user to run the scan.