Slim::Formats: Fix up logging - missing log category
Slim::Formats sets up its log with:

    my $log  = logger('formats');

But the logger category 'formats' does not exist, so it does not show up
as an available logging option.

This change directs logging output to the pre-defined server category
'formats.audio', which is probably most appropriate.

An alternative would be to set up the pre-defined category 'formats'.
But the additional granularity afforded by this approach is unlikely to
be useful.

This issue appears to have been in existence since the adoption of
Log4perl debugging.

Refer commit 6f27c52f 'Implement Log::Log4perl debugging'.