Re-enable update checker in nomysb.com mode.
Previously the check whether a server instance needed an update or not
was run on mysb.com. In order to keep things simple, we're going to rely
on simple, static XML files in the future. Get a version specificy list
of updates, decide in LMS whether we need to update or not.