Unify BMF for volatile and persistent tracks
The mediafolder query doesn't support recursive scans. Therefore we
don't want to use it for BMF playback, as it would disable the feature
where we can play all of eg. an artist's albums stored in sub-folders.
As we can't use mediafolder, adapt the Slim::Utils::Scanner to handle
volatile tracks, too.