Updated fix for crossfade across sample rates
This is an updated fix for temporarily disabling crossfade (if enabled)
between tracks that have different sample rates. As in the original PR
for this, this appears necessary for (at least) the Squeezebox Touch
which will produce an alarming blast of white noise in such a situation.

As reported in PR#3 (https://github.com/Logitech/slimserver/pull/3),
this fix didn't work correctly when near the end of a playlist, for
example when Random Play is adding a track at a time.

The bug was caused because as well has looking between the new track and
previous track to compare for a compatible sample rate, it was actually
looking back even earlier in the playlist for no reason. This was
because I didn't understand the meaning of the offset at this point in
the code.