When we're scanning for new/changed or playlists, playlists which got removed from the disk are not removed from the database.
We did not check for removed items as this caused bug 16037 in the past.
But rather than not to run this check, we should run it selectively:
only compare against list of content type 'ssp' items (internal CT for
playlists). This should prevent the accidental removal of all items,
without ignoring removed playlists.

This change still does not remove tracks referred in the removed
playlist, which are not in the music folder.