Fix web UI's display of artists with a comma in the name.
Don't split "Earth, Wind & Fire". Check whether we have multiple IDs first. Look up individual names in case we end up with more artist names than IDs.

This is not fully fixed in the "Now playing" section of the web UI. But it should work for most cases. Worst case only the first part of the name ("Earth") would be linked, but the full name is shown anyway.