Add pref to control transition across sample rates
It has been reported that not all players have a problem when
crossfading between tracks with a different sample rate and therefore
this adds web and player preference to choose whether to enable or
disable crossfading in this scenario.

This is just the player preference and web settings; it's not wired in
yet.

See here for the request:
https://github.com/Logitech/slimserver/pull/65#issuecomment-226743852