Fix Spotify syncing
Spotify authentication is player specific. Therefore we have to store authentication details in the player object, not the song object. Otherwise playback would fail with invalid credentials.

Whenever a player connects, we make sure it has valid authentication details. Otherwise we might fail again, as when starting the sync we only authenticate for the master player, but not the slaves of a sync group.