Bug 17068: TuneIn doesn't display radio station artwork
Sometimes when a slimservice instances on MySB/UESR is stopped, we might end up with fetchingMeta not being reset. As pluginData is persisted in the database, this would cause a player to never display artwork again. Let's therefore add a timestamp rather than a simple flag, and ignore the timestamp, when it's old.