Don't set up a listener to read from a filehandle of which we know it can't be read.
The debug log output would already have warned about the invalid filehandle. Yet we set it up to be read from. This can lead to a massive number of warnings when using transcoding. Most likely the transcoding pipeline wasn't fully set up at the moment we tried to set up the reader.