Formats::XML - Podcasts: Badly formed RSS image URL breaks Jive
A broken podcast may provide a malformed image url tag. If, for any
reason, we get an empty url tag, or worse, XMLin may interpret it as
a (possibly empty) hash reference, or something else other than a scalar
value. This will break podcast browsing in Jive.

So we add some checks to our retrieved image URLs to ensure that they
are, indeed, scalars. The appended (historic) example is illustrative.

An alternative solution might be available by adapting the parameters
passed to XMLin (in xmlToHash) that control its rendering of the XML.
But that may inadvertently break some other aspect of the parsing.

Example:

Observed in September 2017, although no longer present in the errant
feed. The (edited) feed was as follows:

    <channel>
        <title>A title</title>
        <description>Some description</description>
        <link>A_Link.com</link>
        <pubDate>Sometime in Sept 2017</pubDate>
        <image>
            <url> </url>
            <title>Another title</title>
            <link>Another_Link.com</link>
        </image>
    etc.

If the channel/image/url tag is simply empty, as above, xmlToHash will
return $xml->{'channel'}->{'image'}->{'url'} as an *empty hash ref*,
( image->url = {} ) and *not* a zero length string, or 'undef', as one
might expect.

Jive browsing then breaks when attempting to fetch artwork in
SlimBrowserApplet, because it requires that the icon variable be a
string. It breaks on the structured icon object, even though empty. The
podcast browse menu is not properly rendered, and presents a bewildering
display. Not a good user experience.