If an https stream came without bitrate information, the stream wouldn't play
If there's no bitrate information, LMS would try to start playback, while continuing to read from the stream to be scanned to figure out the bitrate. Unfortunately this seems to somehow lock up streaming when dealing with https (vs. http) sources. In this case delay playback until we're done scanning.