Fix a scanner crash with large collections and dbhighmem enabled.
We would tell SQLite to use memory for temporary tables. This could crash the scanner if a large collection would require a multi hundred megabyte temporary table on eg. a Raspberry Pi with only 1GB of memory. Try to figure out the available memory before running the vacuum, and only use it if there's more than twice the library file's size.