SimpleAsyncHTTP: Don’t cache stale documents - they are never revalidated
Some (older ?) servers may signal a ‘don’t cache’ request by simply returning
the header ‘expires: Thu, 01 Jan 1970 00:00:00 GMT’, without returning any
‘Cache-Control’ headers. Such a document will appear to be permanently stale.
Unwise and very out of date behaviour, perhaps, but it can still be seen in the
wild.

In these circumstances, SimpleAsyncHTTP will seek to cache the document for a
negative period of time ($expires < 0). Unfortunately the effect of this is to
permanently cache the document without ever subjecting it to revalidation.

So, for example, updates to a podcast feed might never be seen unless/until LMS’
cache is cleared. This is not desired behaviour.

Reasons:

 a) The document will never be revalidated, because it passes the
    ‘no revalidaton’ optimization ($expires < $max).
 b) Slim::Utils::Cache/DbCache treats a negative caching period as meaning
    “Never expire”.

The combined effect is that such a document will live permanently in the cache,
and the latest version of the document will never be revalidated/fetched.

This behaviour is not limited to documents served with the header
‘expires: Thu, 01 Jan 1970 00:00:00 GMT’. Any document that appears to be stale
will be so treated. A misconfigured server could produce such a result.

This proposed change deals with the issue by detecting such ‘apparently stale’
documents, and simply not caching them.

Remark: The caller may set an explicit expiration time ($params->{expires}). A
negative value will be ignored.