Formats::XML - Podcasts: Use episode specific image
At present, each episode of a podcast is displayed in the various browse
menus with an accompanying image. This image is taken from the 'channel'
level of the underlying RSS feed, and is identical for each episode. The
standard RSS format makes no provision for images to be provided at the
'item' (episode) level.

But Apple/iTunes extends the format to, optionally, provide a separate
image for each individual episode of a podcast. LMS could make use of
these images, thereby adding a little more variety to the feed display.

This proposed change achieves that.

The episode image, if it exists, is found in the 'href' attribute of an
'itunes:image' tag, at the 'item' level of the feed.

Reference:
https://help.apple.com/itc/podcasts_connect/#/itcb54353390