Slim::Music::Import: Stand-alone scanner - Use running log options
When run as a 'stand-alone' process by Slim::Music::Import::launchScan,
the scanner logs according to:

a) The persistent log settings (those active as a result of "Save
   logging settings for use at next application restart"),

but overridden by:

b) The current log settings actually in force for the scanner log group.

This is in contrast to an 'in-process' scan, which makes use of _all_ of
the current log settings.

So to debug, for example, cue list parsing, it is necessary to set the
desired log level for 'formats.playlists', save log settings "for use at
next application restart", and then run the scanner. After which the
user must remember to put the log levels back to what they were before.

This is not intuitive, and inconsistent with an 'in-process' scan. This
change aligns 'stand-alone' scanner logging behaviour with 'in-process'
scanning.

This is achieved by writing out the current log settings to a transient
log configuration file, and having the stand-alone scanner read its
logging options from that file.

The file is named 'log_transient_scan.conf', and it is placed alongside
the existing persistent log configuration file, 'log.conf', within LMS'
standard file hierarchy. No attempt is made to delete the file when
done, it will simply be overwritten on the next 'stand-alone' scan.