Improve/fix compatibility with Squeezebox Radio
SB Radio firmware unfortunately has a bug failing version string comparison. This leads to LMS8 not being recognized as more recent than 7.7.x. In order to work around this issue we're going to return a fake version string to non-patched Radios. This should allow the user to connect anyway, and install the required patch on the Radio.