add AIF streaming
To complete these patches, just adding AIF as a streamable format. AFAIK, it was just producing white noise before (on my Boom at least).
Only remaining part would be the possibility to strip out the header when streaming proxied sending pcm (not aif/wav). It's not clean to have it work by chance because header is a multiple of samplesize*channels but I really don't think there is a way to fix that as the audio streaming starts before the header check has been done, so we can't skip (or tell the player to skip in case of Direct) ... well this works like that and wav/aif streaming is a corner case anyway :-)

comment typo