Allow plugins to determine how to play URLs
A number of plugins allow you to give them a webpage URL, and they’ll play the music associated with that webpage, e.g., the Band’s Campout plugin has a “Bandcamp URL” field, the YouTube plugin has a “YouTube URL or Video id” field, etc.  However, it would be more convenient if you could paste the URL into the “Tune In URL” field and have LMS work out for you which plugin to use.

This commit provides a mechanism for a plugin to register a regular expression for the URLs it knows how to play.  It can then supply an `explodePlaylist()` method in it’s protocol handler to convert the webpage URL into playable URLs.

See https://github.com/mavit/LMS-YouTube/tree/rich-url-handler for an example plugin adaptation.