Fix #536 An album list filtered by role sometimes doesn't how the role contributor's name
In LMS web UI you can have an album filtered by role `COMPOSER`. Sorting the list by that role could lead to confusing results, as the `COMPOSER` wouldn't be shown as an artist for the albums, but only the album artist.

Make sure we put the role, by which we filter, at the top or the artists list to be shown.