Allow cuesheets to be explicitly tagged as not a cuesheet
The existing compilation functionality would only set if a release was a compilation; a user could not explicitly set that a release was not a compilation.

Additionally, the allowable values for compilation have been extended from compilation values in Slim::Schema->_newTrack. For backward compatiblity, the single character "y" and "n" have been retained