HTTP to HTTPS redir & volatile
This is a tricky one... In general, when a HTTPS subclass was startin a HTTP connection that was then redirected to HTTPS, it was failing (general problem). In addition, for volatile URL when a attempt to re-open() a redirected track fails (gone), then the re-attempt may restart HTTP and then be upgraded. That cannot work in direct mode as players that are HTTP only and have started with HTTP will not be re-directable to HTTPS. So when we re-open on volatile, we'll not use direct if player cannot do HTTPS