Bring the "More" menu item to subscribed podcast feeds.
This required a change in `Slim::Control::XMLBrowser`: if an item has a variable (eg. URL) pointing at a code ref, we must ignore it. These items can't be dealt with by Squeezeplay. In the Podcasts plugin this would eg. be the Search menu item. There's no "More" info menu for such items, which is ok.