songHandler should be currentTrackHandler
This completes the previous work on handlers and trying to clarify the roles of the handlers, between the song/track and the streamUrl.
Previously, the songHandler was prioritized over the streamUrl handler, but it was still missing once case where the $track is a playlist and in fact it's not that $song (master track) handler that should be used but the _currentTrack (sub-track) handler. This is the case where a remote/http playlist is an opml list that contains tracks servced by protocol handlers (e.g. Deezer)