Add $client to a few StreamingController methods
Currently, the Group plugin overwrites Playlist::stopAndClear and Source::playMode just to add the $client (player) requestor of the play/stop/pause/resume. It might not hurt to simply add the $client as a parameter to LMS so that other StreamingController don't have to do the same unelegant overwrite. It's not a big deal though, so feel free to accept it or not. If you do, I'll simply won't load these files when the Group plugin uses LMS 8.2 and above