PodcastIndex cache fix
PodcastIndex requires Unix Epoch time as part of Auth and that caused the http backend cache to almost never be used as request is different every second. Particulary, "new episode" seems to send a different request with different epoch (not sure why) but as a result, we would play a different episode than the selected one as LMS re-issues the request when "play" is pressed. Luckily, PodcastIndex as a 5 minutes window for epoch and this takes advantage of it. But the cache for podcast unfortunately might still need to be re-written