Slim::Formats:XML - RSS - Trim whitespace from Podcast image URL
The URL provided within an RSS image element may contain leading and
trailing white space as well as the text of the URL itself. Generally,
feed publishers do not include leading and trailing white space, but it
need not be so.

LMS does not, at present, trim white space from a URL element. If such
an element is encountered, LMS will fail to retrieve the image due to
the extraneous white space.

This change rectifies matters by applying 'trim' to the retrieved URL.
The RSS parser adopts the same approach when retrieving an audio URL
from the feed.