Podcast: Pre-caching - Guard against fetching same feed multiple times
When browsing within the podcast menu, pre-caching activity will be
initiated for each feed where pre-cached image or podcast info data is
not found.

A guard is needed to prevent launching the pre-cache process multiple
times for the same feed. This will happen whenever the user is browsing
within the podcast menu and a pre-existing pre-cache process has been
launched but has not completed.

This can have an adverse impact on modestly powered servers, and/or
where there are a large number of feeds to be handled.

This change handles the issue by pre-caching place-holder data each time
a feed pre-cache process is started. The place-holder data will be
replaced by real data when the process completes.